requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.0
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.3
backcall==0.2.0
boto3==1.24.36
botocore==1.23.54
cement==2.8.2
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.3
comm==0.1.4
contourpy==1.1.0
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==0.5.0
Django==5.1.3
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-allauth==0.51.0
django-ckeditor==6.5.1
django-cors-headers==4.6.0
django-excel==0.0.10
django-honeypot==1.0.3
django-import-export==3.2.0
django-js-asset==2.0.0
# Editable install with no version control (django-jsignature==0.11)
-e c:\users\user\downloads\django-jsignature-master\django-jsignature-master
django-mathfilters==1.0.0
django-multipleselectfield==0.2.0.post2
django-multiselectfield==0.1.12
django-qr-code==3.1.1
django-qrcode==0.3
django-session-timeout==0.1.0
django-storages==1.12.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ebcli==4.0.18
enum34==1.1.10
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.0
frozenlist==1.4.0
future==0.16.0
geographiclib==2.0
geopy==2.4.0
idna==3.3
import-export==0.3.1
ipykernel==6.25.2
ipython==8.15.0
jaraco.classes==3.2.3
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
jupyter_client==8.3.1
jupyter_core==5.3.1
keyring==9.0
keyrings.alt==4.2.0
kiwisolver==1.4.4
lml==0.1.0
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==8.14.0
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.25.2
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
Pygments==2.16.1
PyJWT==2.4.0
pyparsing==3.0.9
pypiwin32==223
pypng==0.20220715.0
pyproj==3.6.1
pyquery==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2020.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==25.1.1
qrcode==7.4.2
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.6.0
scipy==1.11.3
segno==1.6.0
semantic-version==2.8.5
simplejson==3.19.1
six==1.14.0
sqlparse==0.3.1
stack-data==0.6.2
tablib==3.5.0
termcolor==1.1.0
texttable==1.6.4
tornado==6.3.3
traitlets==5.10.0
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.1.9
whitenoise==5.1.0
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
yarl==1.9.2
